bug fix: int64 cast properly

+++ dcos/resource_dcos_job.go
@@ -428,7 +428,7 @@ func generateMetronomeJob(d *schema.ResourceData, meta interface{}) dcos.Metrono
 	}
 
 	if disk, ok := d.GetOk("disk"); ok {
-		metronome_job_run.Disk = disk.(int64)
+		metronome_job_run.Disk = int64(disk.(int))
 	}
 
 	// labels
